Changes pg_dump operations to copying logs Changes pg_dump operations to copying files
The intent here is to create a postgres log exporter to S3.  There
already existed this project to periodically pg_dump and stream
those to S3 so using that.

This will look in a mounted folder, /pglogs, for a logfile
from yesterday, ie "postgresql-Thu.log" and then "cat" the file
streaming it into S3.  This could be modified to stream the current
logfile if so inclined, but if there is a future iteration of this
it will probably just use a bash script and copy yesterday's logfile
versus streaming it.